enhance(apps/frontend-manage): add setting to change project update email frequency
