+++ packages/graphql/src/public/client.json
@@ -6,6 +6,7 @@
    "ChangeCourseColor": "3d483bc4021ae8f471409fb6d1994f3e00678595a665c15359c7b5b9b79d89d0",
    "ChangeCourseDates": "f4f10126d929ae575e7e4a416611c1608b14360215461d4c4733888b136f5958",
    "ChangeCourseDescription": "b11fa59c1381e2682eb55a8f06436a0eff89006c0c31eaf3db05472b888c616d",
+   "ChangeEmailSettings": "4eb1aea4a048c1f61c892d0efe44e45ce29c2f45066a91d1bebfeae593ff6fa3",
    "ChangeInitialSettings": "b7910286bae558ed5213ef143aff669028298f7ced2f7e8ec8dabbeecdf4453f",
    "ChangeParticipantLocale": "a35b384c5614ee4c314767f21d4310c0b1be94ef14c6759ef275a4bea5efbba3",
    "ChangeSessionSettings": "a96f0ba9beeae844e8db86511cdc85cf00d71679116880554a24e23415b1d627",
@@ -121,7 +122,7 @@
    "Participations": "88903a59853ede963d9369b0a9ceb43adaa308b7bcb527c5682dd8b73c96a419",
    "Self": "05603b4fdfd68bbe0990c57cb9e59a5f71ef9705d23d928fd63589dd5364f945",
    "SelfWithAchievements": "f161d562f384a81dbecd1baa222201522702925a82f859a1238038fbd35413fb",
-   "UserProfile": "116fb85886de9ee42ce8776fb77d65209056cde3115e104124761043d236697d",
+   "UserProfile": "d37d1882bef914e250cba85de86301e48ed8fca440cde15f3568320219b11e1c",
    "FeedbackAdded": "b7948b93f7418778d06c2b3a18d84b3a46c071d3beb1c539aa15d727d327c264",
    "FeedbackCreated": "1447a6241024f772cef8129c6676933db63ab227a0744e9e2ff73a024d838759",
    "FeedbackRemoved": "483f0c5d7011474da36ec8feb384ec2f9e5877a4283f75f6718da405e616bab6",
