+++ packages/graphql/src/public/server.json
@@ -6,6 +6,7 @@
    "3d483bc4021ae8f471409fb6d1994f3e00678595a665c15359c7b5b9b79d89d0": "mutation ChangeCourseColor($color: String!, $courseId: String!) {\n  changeCourseColor(color: $color, courseId: $courseId) {\n    id\n    color\n    __typename\n  }\n}",
    "f4f10126d929ae575e7e4a416611c1608b14360215461d4c4733888b136f5958": "mutation ChangeCourseDates($courseId: String!, $startDate: Date, $endDate: Date) {\n  changeCourseDates(courseId: $courseId, startDate: $startDate, endDate: $endDate) {\n    id\n    startDate\n    endDate\n    __typename\n  }\n}",
    "b11fa59c1381e2682eb55a8f06436a0eff89006c0c31eaf3db05472b888c616d": "mutation ChangeCourseDescription($input: String!, $courseId: String!) {\n  changeCourseDescription(input: $input, courseId: $courseId) {\n    id\n    description\n    __typename\n  }\n}",
+   "4eb1aea4a048c1f61c892d0efe44e45ce29c2f45066a91d1bebfeae593ff6fa3": "mutation ChangeEmailSettings($projectUpdates: Boolean!) {\n  changeEmailSettings(projectUpdates: $projectUpdates) {\n    id\n    sendProjectUpdates\n    __typename\n  }\n}",
    "b7910286bae558ed5213ef143aff669028298f7ced2f7e8ec8dabbeecdf4453f": "mutation ChangeInitialSettings($shortname: String!, $locale: LocaleType!) {\n  changeInitialSettings(shortname: $shortname, locale: $locale) {\n    id\n    email\n    shortname\n    locale\n    firstLogin\n    catalyst\n    catalystTier\n    __typename\n  }\n}",
    "a35b384c5614ee4c314767f21d4310c0b1be94ef14c6759ef275a4bea5efbba3": "mutation ChangeParticipantLocale($locale: LocaleType!) {\n  changeParticipantLocale(locale: $locale) {\n    id\n    locale\n    __typename\n  }\n}",
    "a96f0ba9beeae844e8db86511cdc85cf00d71679116880554a24e23415b1d627": "mutation ChangeSessionSettings($id: String!, $isLiveQAEnabled: Boolean, $isConfusionFeedbackEnabled: Boolean, $isModerationEnabled: Boolean, $isGamificationEnabled: Boolean) {\n  changeSessionSettings(id: $id, isLiveQAEnabled: $isLiveQAEnabled, isConfusionFeedbackEnabled: $isConfusionFeedbackEnabled, isModerationEnabled: $isModerationEnabled, isGamificationEnabled: $isGamificationEnabled) {\n    id\n    isLiveQAEnabled\n    isConfusionFeedbackEnabled\n    isModerationEnabled\n    isGamificationEnabled\n    __typename\n  }\n}",
@@ -121,7 +122,7 @@
    "88903a59853ede963d9369b0a9ceb43adaa308b7bcb527c5682dd8b73c96a419": "query Participations($endpoint: String) {\n  participations(endpoint: $endpoint) {\n    id\n    completedMicroSessions\n    subscriptions {\n      id\n      endpoint\n      __typename\n    }\n    course {\n      id\n      displayName\n      startDate\n      endDate\n      description\n      isGamificationEnabled\n      microSessions {\n        id\n        displayName\n        scheduledStartAt\n        scheduledEndAt\n        __typename\n      }\n      sessions {\n        id\n        displayName\n        linkTo\n        __typename\n      }\n      __typename\n    }\n    __typename\n  }\n}",
    "05603b4fdfd68bbe0990c57cb9e59a5f71ef9705d23d928fd63589dd5364f945": "query Self {\n  self {\n    id\n    email\n    username\n    avatar\n    avatarSettings\n    isActive\n    isProfilePublic\n    xp\n    level\n    levelData {\n      id\n      index\n      name\n      avatar\n      requiredXp\n      nextLevel {\n        id\n        index\n        requiredXp\n        avatar\n        name\n        __typename\n      }\n      __typename\n    }\n    __typename\n  }\n}",
    "f161d562f384a81dbecd1baa222201522702925a82f859a1238038fbd35413fb": "query SelfWithAchievements {\n  selfWithAchievements {\n    participant {\n      id\n      username\n      avatar\n      avatarSettings\n      xp\n      level\n      levelData {\n        id\n        index\n        name\n        avatar\n        requiredXp\n        nextLevel {\n          id\n          index\n          requiredXp\n          avatar\n          name\n          __typename\n        }\n        __typename\n      }\n      achievements {\n        id\n        achievedAt\n        achievedCount\n        achievement {\n          id\n          nameDE\n          nameEN\n          descriptionDE\n          descriptionEN\n          icon\n          iconColor\n          __typename\n        }\n        __typename\n      }\n      __typename\n    }\n    achievements {\n      id\n      nameDE\n      nameEN\n      descriptionDE\n      descriptionEN\n      icon\n      iconColor\n      __typename\n    }\n    __typename\n  }\n}",
-   "116fb85886de9ee42ce8776fb77d65209056cde3115e104124761043d236697d": "query UserProfile {\n  userProfile {\n    id\n    email\n    shortname\n    locale\n    firstLogin\n    catalyst\n    catalystTier\n    __typename\n  }\n}",
+   "d37d1882bef914e250cba85de86301e48ed8fca440cde15f3568320219b11e1c": "query UserProfile {\n  userProfile {\n    id\n    email\n    sendProjectUpdates\n    shortname\n    locale\n    firstLogin\n    catalyst\n    catalystTier\n    __typename\n  }\n}",
    "b7948b93f7418778d06c2b3a18d84b3a46c071d3beb1c539aa15d727d327c264": "fragment FeedbackData on Feedback {\n  id\n  isPublished\n  isPinned\n  isResolved\n  content\n  votes\n  resolvedAt\n  createdAt\n  responses {\n    id\n    content\n    positiveReactions\n    negativeReactions\n    __typename\n  }\n  __typename\n}\nsubscription FeedbackAdded($sessionId: String!) {\n  feedbackAdded(sessionId: $sessionId) {\n    ...FeedbackData\n    __typename\n  }\n}",
    "1447a6241024f772cef8129c6676933db63ab227a0744e9e2ff73a024d838759": "fragment FeedbackData on Feedback {\n  id\n  isPublished\n  isPinned\n  isResolved\n  content\n  votes\n  resolvedAt\n  createdAt\n  responses {\n    id\n    content\n    positiveReactions\n    negativeReactions\n    __typename\n  }\n  __typename\n}\nsubscription FeedbackCreated($sessionId: String!) {\n  feedbackCreated(sessionId: $sessionId) {\n    ...FeedbackData\n    __typename\n  }\n}",
    "483f0c5d7011474da36ec8feb384ec2f9e5877a4283f75f6718da405e616bab6": "subscription FeedbackRemoved($sessionId: String!) {\n  feedbackRemoved(sessionId: $sessionId)\n}",
