+++ packages/graphql/src/schema/mutation.ts
@@ -912,6 +912,17 @@ export const Mutation = builder.mutationType({
         },
       }),
 
+      changeEmailSettings: t.withAuth(asUserFullAccess).field({
+        nullable: true,
+        type: User,
+        args: {
+          projectUpdates: t.arg.boolean({ required: true }),
+        },
+        resolve(_, args, ctx) {
+          return AccountService.changeEmailSettings(args, ctx)
+        },
+      }),
+
       changeInitialSettings: t.withAuth(asUserFullAccess).field({
         nullable: true,
         type: User,
