+++ packages/i18n/messages/de.ts
@@ -669,6 +669,9 @@ Da die KlickerUZH-App noch nicht im iOS-App-Store verfügbar ist, folgen Sie die
       shortnameAlphanumeric:
         'Der Kurzname darf nur aus Buchstaben und Zahlen bestehen.',
       shortnameTaken: 'Dieser Kurzname ist bereits vergeben.',
+      emailUpdates: 'E-Mail Updates',
+      emailUpdatesTooltip:
+        'Diese Einstellung beeinflusst die E-Mails, die Sie in Verbindung mit KlickerUZH erhalten. E-Mails zu wichtigen Releases werden immer an Ihre Edu-ID E-Mail-Adresse gesendet, häufigere Projekt-Updates können hier aktiviert oder deaktiviert werden.',
     },
     token: {
       pageName: 'Token Generation',
